llm-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-genai>=0.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0
pytest>=8.0.0
tzdata>=2024.1
bs4>=0.0.2